pyproject.toml
[project]
name = "promptly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "anthropic>=0.57.1",
    "fsspec>=2025.9.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "mistralai>=1.9.2",
    "numpy>=2.2.6",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "setuptools>=80.9.0",
]
